Fix divide-by-zero bug in HSL/HSV conversion
When calculating the hue, the zero case needs to be checked first to
prevent a divide-by-zero.  Note that this could be done using
`max' == 0` as in the calculation of `s`, but I think that explicitly
defining the chroma (`c'`) is more clear, and the value is used more
than once in the HSV case regardless.